requirements.txt
docopt~=0.6.2
PyGitHub~=1.45
